revert: "fix: Make build compatible with new GDAL container TDE-1179 (#971)" (#975)
#### Motivation

This reverts commit 719d96e1ff1133c351bbee7a1d93fa359cc702e4.
We are not using GDAL 3.9 and we had issues running the container within
Argo Workflows since this change that modified how the Python
environment is managed.
There are some jobs that need to be run on Argo workflows this week so
we need to allow user to do it while we are finding a fix for the Argo
issue.

#### Modification

_Why is this change being made? What implications or other
considerations are there?_

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title