feat: Add created/updated properties to collections TDE-1147 (#981)
These fields are not part of core STAC, and are not yet in the TTW LINZ
extension. This has been relegated to a separate story
<https://toitutewhenua.atlassian.net/browse/TDE-1192>.

#### Motivation

Make it easier for end users to determine when a dataset was
created/updated by providing these properties in the top level STAC
document.

#### Checklist

- [x] Tests updated
- [ ] Docs updated (N/A; `collection.json` structure is not documented
in this repository)
- [x] Issue linked in Title