fix: concurrency file order TDE-1213 (#998)
#### Motivation

When supplying multiple source locations to use for standardising, if
there are overlapping source images, the VRT created does not honour the
priority order for the supplied datasets. This Pull Request is to retain
the priority of the images for merging.

#### Modification

Move away from using `as_completed` for parallel processing so that the
output file list is in the same order as the input file list.

#### Checklist

_If not applicable, provide explanation of why._

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title