ci: publish container to gh on pr label container TDE-1252 (#1038)
### Motivation

It is useful to be able to publish a container from a PR code base so we
can test the container using external environment such as Kubernetes.

### Modifications

- GH Actions workflow `containers.yml` will run when a `container` label
is added to a PR. The container will only be published on the GH package
repository and be tagged `pr-[PR-NUMBER]`

### Verification

You can test it in this PR.
Also tested on a fork with the full process of merging the PR for non
regression.

---------

Co-authored-by: Alice Fage <afage@linz.govt.nz>