test: use equals_exact to check winding order of capture area (#1062)
### Motivation

Because the test were using `equals` they were checking that the shapes
were equivalent, but not exactly the same (i.e. including the winding
order). In fact [due to a change in
orientation](https://toitutewhenua.atlassian.net/browse/TDE-1205), they
were reversed, but `equals` does not check orientation.


![winding_order](https://github.com/user-attachments/assets/40b8edd7-043d-4f2c-a85b-87a0c6a9a881)

This change is to reflect the correct data in order and use
`equals_exact` to catch if this changes in future.

### Modifications

Use Shapely `Polygon.equals_exact()` instead of `Polygon.equals()` and
update with the correct capture area test data.

### Verification

pytest unit tests.