fix: Don't attempt to install project inside virtualenv (#1074)
### Motivation

Avoids the following:

```
Installing the current project: […]
Warning: The current project could not be installed: No file/folder found for package […]
If you do not want to install the current project use --no-root.
If you want to use Poetry only for dependency management but not for packaging, you can disable package mode by setting package-mode = false in your pyproject.toml file.
In a future version of Poetry this warning will become an error!
```

### Verification

N/A