feat!: add linz:slug to Collection TDE-1305 (#1152)
### Motivation

As a Basemaps developer I want the dataset name to be referenced within
the collection.json so that I can look this up and use it as part of a
link to the same dataset in Basemaps.

### Modifications

- Add `--linz-slug` argument to `collection_from_items.py`
- Add `linz:slug` attribute to `stac` property of `ImageryCollection`
- Update `ImageryCollection` class signature
- Update `create_collection` function signature
- Pass `linz_slug` from `collection_from_items.py` to
`ImageryCollection`
- Tests updated

### Verification

Updated pytest tests to include `linz:slug` (and failure when the
respective argument is missing).
Additional manual testing.