revert: normalize capture-area geometry TDE-1359 (#1288) TDE-1423 (#1292)
This reverts commit 0a061ba3910a2ed60aecf8b9ad1ec5ebc5cf80df.

### Motivation

Normalizing as it has been implemented brings some issues into the
overall geometry (visual and QGIS artifacts). As discussed and decided,
having the capture-area content being different (points ordering) that
shows a difference when updating an existing dataset while its geometry
is not visually changed should not be a big deal for data consumer.

### Modifications

- Revert https://github.com/linz/topo-imagery/pull/1288

### Verification

None