fix: handle optional flags from Argo TDE-1346 (#1287)
### Motivation
Argo Workflows supplies single quotes where parameters are not supplied,
which can happen if a workflow template is used by multiple workflows.
This breaks optional parameters in argparse.

### Modifications

Convert Argo string to boolean to control adding the capture dates file
and the title suffix, in line with how the capture area creation
(footprint) is handled.

### Verification

Ran workflow in Argo and locally on the command line.