fix: existing collection is not updated with new metadata TDE-1444 (#1313)
### Motivation

When updating a dataset (added/modified TIFFs for example), the new
metadata that is passed to the `collection_from_items.py` script is not
used to update the existing Collection.

### Modifications

- Update the existing Collection metadata from the new metadata passed
to the `collection_from_items.py` script
- Allow to bypass the `title` re-generation using `--keep-title` flag

>__Note__: If an existing optional metadata field (for example,
`--event`) is not passed to the script, the existing metdata field (for
example, `linz:event_name`) will be removed.

<!-- TODO: Attach screenshots if you changed the UI. -->
Example for National DEM:

![image](https://github.com/user-attachments/assets/7af65804-03b6-414a-9ad8-c0f16232c8e7)

Example not passing the same metadata:

![image](https://github.com/user-attachments/assets/f9a3dca7-e6bd-4f52-bae3-c6cd9cadff19)

### Verification

- Automated tests
- Argo Workflows