feat: add GSD metadata to collection.json TDE-1657 (#1407)
### Motivation
As an imagery maintainer
I want to make sure the GSD values across a survey is consistent in all
TIFF images
so that the data is consistent and accurate (in line with the metadata)

GSD value of the associated assets needs to be saved in the
`collection.json` file for easy access when validating GSD in
stac-setup.


### Modifications
Added `gsd` metadata to our STAC collection.

### Verification

added unit tests.