fix: gdalbuildvrt - specify EPSG code to prevent complex VRT cutline geometry (#1454)
### Motivation

When standardising imagery where the projection information is not
stored in the TIFF file, a complex geometry is created in the VRT when
applying a cutline using `gdalwarp`, and the following warning is
logged:

`Warning 1: the input vector layer has a SRS, but the source raster
dataset does not. Cutline results may be incorrect.`

This causes the `gdal_translate` command downstream to take a very long
time (hours) to complete, particularly when retiling.

### Modifications

Specify the EPSG code in the `gdalbuildvrt` command using the value
supplied in the `--source-epsg` workflow parameter. Default to EPSG:2193
if not specified (when used by the hillshade workflows). This prevents
the downstream issues with `gdalwarp` and `gdal_translate`.

The `gdal_translate` command already uses `--source-epsg` so this change
does not change the fundamental EPSG code logic.

### Verification

Local testing and workflow run against PR container.