refactor: remove unused paletted bands checks TDE-1753 (#1458)
### Motivation

The removed code was created in
[PR#269](https://github.com/linz/topo-imagery/pull/269) but was
superseded by changes in
[PR#547](https://github.com/linz/topo-imagery/pull/547). This was not
detected due to a misspelling of the Palette band in the test.

### Modifications

Removed defaults that were used for greyscale paletted TIFFs prior to
handling colour paletted TIFFs. Paletted TIFFs will always be
standardised as RGB(A) as we cannot tell if they are greyscale from the
bands' information.

Throw an error if RGB imagery does not contain Red, Green and Blue
colorInterpretation.

### Verification

Unit tests.